JetBrains: Cody: Split settings for plugin modules
