+++ client/jetbrains/src/main/java/com/sourcegraph/config/ConfigUtil.java
@@ -24,6 +24,9 @@ import org.jetbrains.annotations.Nullable;
 public class ConfigUtil {
   public static final String DOTCOM_URL = "https://sourcegraph.com/";
   public static final String SERVICE_DISPLAY_NAME = "Sourcegraph Cody + Code Search";
+  public static final String CODY_DISPLAY_NAME = "Cody AI";
+  public static final String CODE_SEARCH_DISPLAY_NAME = "Code Search";
+  public static final String SOURCEGRAPH_DISPLAY_NAME = "Sourcegraph";
 
   @NotNull
   public static ExtensionConfiguration getAgentConfiguration(@NotNull Project project) {
