+++ client/jetbrains/src/main/java/com/sourcegraph/config/OpenPluginSettingsAction.java
@@ -4,7 +4,7 @@ import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.options.ShowSettingsUtil;
 import com.intellij.openapi.project.DumbAwareAction;
 import com.intellij.openapi.util.NlsActions;
-import com.sourcegraph.cody.config.SettingsConfigurable;
+import com.sourcegraph.cody.config.ui.AccountConfigurable;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -20,6 +20,6 @@ public class OpenPluginSettingsAction extends DumbAwareAction {
   @Override
   public void actionPerformed(@NotNull AnActionEvent event) {
     ShowSettingsUtil.getInstance()
-        .showSettingsDialog(event.getProject(), SettingsConfigurable.class);
+        .showSettingsDialog(event.getProject(), AccountConfigurable.class);
   }
 }
