+++ client/jetbrains/src/main/java/com/sourcegraph/telemetry/PostStartupActivity.java
@@ -7,7 +7,8 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.startup.StartupActivity;
 import com.sourcegraph.cody.CodyAgentProjectListener;
 import com.sourcegraph.cody.config.CodyApplicationSettings;
-import com.sourcegraph.config.SettingsChangeListener;
+import com.sourcegraph.cody.config.notification.AccountSettingChangeListener;
+import com.sourcegraph.cody.config.notification.CodySettingChangeListener;
 import java.util.UUID;
 import org.jetbrains.annotations.NotNull;
 
@@ -18,8 +19,9 @@ public class PostStartupActivity implements StartupActivity.DumbAware {
 
   @Override
   public void runActivity(@NotNull Project project) {
-    // Make sure that SettingsChangeListener is loaded
-    project.getService(SettingsChangeListener.class);
+    // Make sure that settings -ChangeListeners are loaded
+    project.getService(AccountSettingChangeListener.class);
+    project.getService(CodySettingChangeListener.class);
 
     // When no anonymous user ID is set yet, we create a new one and treat this as an installation
     // event.
