+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/notification/AccountSettingChangeActionNotifier.kt
@@ -1,0 +1,16 @@
+package com.sourcegraph.cody.config.notification
+
+import com.intellij.util.messages.Topic
+
+interface AccountSettingChangeActionNotifier {
+    companion object {
+        @JvmStatic
+        val TOPIC = Topic.create(
+            "Sourcegraph Cody + Code Search plugin settings have changed",
+            AccountSettingChangeActionNotifier::class.java)
+    }
+
+    fun beforeAction(context: AccountSettingChangeContext)
+
+    fun afterAction(context: AccountSettingChangeContext)
+}
