+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/notification/AccountSettingChangeContext.kt
@@ -1,0 +1,3 @@
+package com.sourcegraph.cody.config.notification
+
+class AccountSettingChangeContext(val serverUrlChanged: Boolean = false, val accessTokenChanged: Boolean = false)
