+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/notification/CodySettingChangeActionNotifier.kt
@@ -1,0 +1,16 @@
+package com.sourcegraph.cody.config.notification
+
+import com.intellij.util.messages.Topic
+
+interface CodySettingChangeActionNotifier {
+    companion object {
+        @JvmStatic
+        val TOPIC = Topic.create(
+            "Sourcegraph Cody + Code Search: Cody AI settings have changed",
+            CodySettingChangeActionNotifier::class.java)
+    }
+
+    fun beforeAction(context: CodySettingChangeContext)
+
+    fun afterAction(context: CodySettingChangeContext)
+}
