+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/notification/CodySettingChangeContext.kt
@@ -1,0 +1,12 @@
+package com.sourcegraph.cody.config.notification
+
+class CodySettingChangeContext(val oldCodyEnabled: Boolean,
+                               val newCodyEnabled: Boolean,
+                               val oldCodyAutocompleteEnabled: Boolean,
+                               val newCodyAutocompleteEnabled: Boolean,
+                               val oldIsCustomAutocompleteColorEnabled: Boolean,
+                               val isCustomAutocompleteColorEnabled: Boolean,
+                               val oldCustomAutocompleteColor: Int?,
+                               val customAutocompleteColor: Int?,
+                               val oldBlacklistedAutocompleteLanguageIds: List<String>,
+                               val newBlacklistedAutocompleteLanguageIds: List<String>)
