+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/notification/CodySettingChangeListener.kt
@@ -1,0 +1,102 @@
+package com.sourcegraph.cody.config.notification
+
+import com.intellij.openapi.Disposable
+import com.intellij.openapi.diagnostic.Logger
+import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.wm.ToolWindowManager
+import com.intellij.util.messages.MessageBusConnection
+import com.sourcegraph.cody.CodyAgentProjectListener
+import com.sourcegraph.cody.CodyToolWindowFactory
+import com.sourcegraph.cody.agent.CodyAgent
+import com.sourcegraph.cody.autocomplete.CodyAutocompleteManager
+import com.sourcegraph.cody.autocomplete.render.AutocompleteRenderUtils
+import com.sourcegraph.cody.config.CodyApplicationSettings
+import com.sourcegraph.cody.statusbar.CodyAutocompleteStatus
+import com.sourcegraph.cody.statusbar.CodyAutocompleteStatusService.Companion.notifyApplication
+import com.sourcegraph.config.ConfigUtil
+import com.sourcegraph.find.browser.JavaToJSBridge
+import com.sourcegraph.utils.CollectionUtil.Companion.diff
+import java.util.function.Consumer
+
+
+class CodySettingChangeListener(val project: Project) : Disposable {
+    var connection: MessageBusConnection? = null
+    var javaToJSBridge: JavaToJSBridge? = null
+    val logger = Logger.getInstance(CodySettingChangeListener::class.java)
+
+    init {
+        val bus = project.messageBus
+        connection = bus.connect()
+        connection?.subscribe(
+            CodySettingChangeActionNotifier.TOPIC,
+            object : CodySettingChangeActionNotifier {
+                override fun beforeAction(context: CodySettingChangeContext) {
+                }
+
+                override fun afterAction(context: CodySettingChangeContext) {
+                    // Notify JCEF about the config changes
+                    javaToJSBridge?.callJS("pluginSettingsChanged", ConfigUtil.getConfigAsJson(project))
+
+                    // Notify Cody Agent about config changes.
+                    val agentServer = CodyAgent.getServer(project)
+                    if (ConfigUtil.isCodyEnabled() && agentServer != null) {
+                        agentServer.configurationDidChange(ConfigUtil.getAgentConfiguration(project))
+                    }
+
+                    if (context.newCodyEnabled) {
+                        // Starting the agent is idempotent, so it's OK if we call startAgent multiple times.
+                        CodyAgentProjectListener.startAgent(project)
+                    } else {
+                        // Stopping the agent is idempotent, so it's OK if we call stopAgent multiple times.
+                        CodyAgentProjectListener.stopAgent(project)
+                    }
+
+                    // clear autocomplete suggestions if freshly disabled
+                    if (context.oldCodyAutocompleteEnabled && !context.newCodyAutocompleteEnabled) {
+                        CodyAutocompleteManager.getInstance().clearAutocompleteSuggestionsForAllProjects()
+                    }
+
+                    // Disable/enable the Cody tool window depending on the setting
+                    if (!context.newCodyEnabled && context.oldCodyEnabled) {
+                        val toolWindowManager = ToolWindowManager.getInstance(project)
+                        val toolWindow = toolWindowManager.getToolWindow(CodyToolWindowFactory.TOOL_WINDOW_ID)
+                        toolWindow?.setAvailable(false, null)
+                    } else if (context.newCodyEnabled && !context.oldCodyEnabled) {
+                        val toolWindowManager = ToolWindowManager.getInstance(project)
+                        val toolWindow = toolWindowManager.getToolWindow(CodyToolWindowFactory.TOOL_WINDOW_ID)
+                        toolWindow?.setAvailable(true, null)
+                    }
+                    if (!context.newCodyEnabled) {
+                        notifyApplication(CodyAutocompleteStatus.CodyDisabled)
+                    } else if (!context.newCodyAutocompleteEnabled) {
+                        notifyApplication(
+                            CodyAutocompleteStatus.AutocompleteDisabled)
+                    } else {
+                        notifyApplication(CodyAutocompleteStatus.Ready)
+                    }
+
+                    // Rerender autocompletions when custom autocomplete color changed
+                    // or when checkbox state changed
+                    if (context.oldCustomAutocompleteColor != context.customAutocompleteColor
+                        || (context.oldIsCustomAutocompleteColorEnabled
+                            != context.isCustomAutocompleteColorEnabled)) {
+                        ConfigUtil.getAllEditors()
+                            .forEach(Consumer { editor: Editor? -> AutocompleteRenderUtils.rerenderAllAutocompleteInlays(editor) })
+                    }
+
+                    // clear autocomplete inlays for blacklisted language editors
+                    val languageIdsToClear: List<String> =
+                        context.newBlacklistedAutocompleteLanguageIds
+                            .diff(context.oldBlacklistedAutocompleteLanguageIds)
+                    if (languageIdsToClear.isNotEmpty())
+                        CodyAutocompleteManager.getInstance()
+                            .clearAutocompleteSuggestionsForLanguageIds(languageIdsToClear)
+                }
+            })
+    }
+
+    override fun dispose() {
+        connection!!.disconnect()
+    }
+}
